\name{RockFab-package}
\alias{RockFab-package}
\alias{RockFab}
\docType{package}
\title{
Rock fabric and strain analysis tools.
}
\description{
Provides functions to complete three-dimensional rock fabric and strain analyses following the Rf Phi, Fry, and normalized Fry methods. Also allows for plotting of results and interactive 3D visualization functionality.
}
\details{
\tabular{ll}{
Package: \tab RockFab\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-02\cr
License: \tab GPL\cr
}

}
\author{
Jeffrey R. Webber

Maintainer: Jeffrey R. Webber <jwebber@geo.umass.edu>
}
